add intro to lvl 2
