+++ app.js
@@ -1,7 +1,5 @@
-import { soundMainTheme, soundIntro} from "./music.js";
-import { startGame, endGameAction } from './gameManager.js';
-
-let isGameStarted = false;
+import { soundIntro} from "./music.js";
+import { endGameAction } from './gameManager.js';
 
 function handleStartGameBtnClick() {
   soundIntro.play();
@@ -13,22 +11,7 @@ function hideStartScreen() {
   startScreen.remove();
 }
 
-let introduction = document.querySelector('#introduction');
 const startGameButton = document.querySelector("#startGameButton");
 const endGameButton = document.querySelector('#endGameButton');
 endGameButton.addEventListener('click', endGameAction)
 startGameButton.addEventListener("click", handleStartGameBtnClick);
-
-function introductionSkip() {
-  if (!isGameStarted) {
-    startGame();
-    soundMainTheme.play();
-    soundMainTheme.volume = 0.3;
-    introduction.remove();
-    introduction.style.zIndex = 1;
-    soundIntro.pause();
-    isGameStarted = true;
-  } 
-}
-
-introduction.addEventListener('click', introductionSkip);
