+++ enemy.js
@@ -58,7 +58,7 @@ function createEnemy(src, width, height, i, maxHealth) {
 }
 
 export function createEnemies() {
-  for(let i = 0; i < 11; i++) {
+  for(let i = 0; i < 1; i++) {
     let enemy;  
     if(i < 5){
       enemy = createEnemy('img/su-3.png', 250, 80, i, 50)
