+++ gameManager.js
@@ -1,9 +1,13 @@
 import { mobileControls } from './touch.js';
 import { rocket } from './rocket.js';
 import { airfighter } from "./airfighter.js";
-import { deleteEnemies, enemies, createEnemies } from './enemy.js';
-import { soundRocketShot } from "./music.js";
+import { deleteEnemies, createEnemies } from './enemy.js';
+import { soundRocketShot, soundMainTheme, soundIntro } from "./music.js";
 import { Step, gameState, fps } from './step.js';
+
+let isGameStarted = false;
+let introduction = document.querySelector('#introduction');
+introduction.addEventListener('click', introductionSkip);
 
 export const levelState = {
   isLevelFinished : false,
@@ -11,20 +15,44 @@ export const levelState = {
 }
 
 export function startGame() {
-    initKeybordMovement();
+  console.log('startGame');
+  initKeybordMovement();
+  if (levelState.levelNumber == 1) {
     startLevel1()
+  }
+  if (levelState.levelNumber == 2) {
+    startLevel2()
+  }
+}
+
+function introductionSkip() {
+  console.log('funtion introductionSkip');
+  if (!isGameStarted) {
+    console.log('funtion introductionSkip: Starting Game');
+    startGame();
+    soundMainTheme.play();
+    soundMainTheme.volume = 0.3;
+    introduction.style.display = "none";
+    introduction.style.zIndex = 1;
+    soundIntro.pause();
+    isGameStarted = true;
+  } 
 }
 
 function startLevel1() {
+  console.log('startLevel1');
   gameState.gameIntervalId = setInterval(Step, 1000 / fps);
 }
 
 function startLevel2() {
+  isGameStarted = false;
+  console.log('startLevel2');
+  levelState.levelNumber = 2;
   deleteEnemies();
   createEnemies()
   gameState.gameIntervalId = setInterval(Step, 1000 / fps);
   document.querySelector('#levelComplete').style.display = 'none'
-
+  introduction.style.display = "block";
 }
 
 document.querySelector('#nextlevel').addEventListener('click', startLevel2)
